fix(lib): revert ARexx function return type change
Original commit: 8f4987da4f8b98a45c5e5523c598744a62ad77ae

This changed the API of DICE - we don't want to break external code like
this. Instead we want to keep the API as it was and adapt the
implementation to be compatible.

+++ include/lib/rexx.h
@@ -10,8 +10,6 @@
 #ifndef LIB_REXX_H
 #define LIB_REXX_H
 
-#include <stdint.h>
-
 #ifndef EXEC_TYPES_H
 #include <exec/types.h>
 #endif
@@ -19,8 +17,8 @@
 #include <exec/ports.h>
 #endif
 
-int32_t PlaceRexxCommand(struct MsgPort *, char *, char **, int32_t *);
-int32_t PlaceRexxCommandDirect(struct MsgPort *, char *, char *, char **, int32_t *);
+long PlaceRexxCommand(struct MsgPort *, char *, char **, long *);
+long PlaceRexxCommandDirect(struct MsgPort *, char *, char *, char **, long *);
 void ProcessRexxCommands(struct MsgPort *);
 short CreateGlobalDiceRexxPort(struct MsgPort *, char *);
 short CreateDiceRexxPort(struct MsgPort *, char *);
@@ -46,7 +44,7 @@ extern struct Library *RexxSysBase; /*  NULL if DICE was unable to open lib*/
  *  message ports) when applicable, -1 on error.
  */
 
-int32_t DoRexxCommand(void *rexxmsg, struct MsgPort *port, char *arg0, char **resstr);
+long DoRexxCommand(void *rexxmsg, struct MsgPort *port, char *arg0, char **resstr);
 
 extern char *RexxHostName;
 extern short RexxSigBit;
